Move into BossBot and add ChatBot's go.mod module description
