+++ ChatBot/go.mod
@@ -1,0 +1,6 @@
+module ChatBot
+
+require (
+	github.com/pkg/errors v0.8.0
+	github.com/sirupsen/logrus v1.2.0
+)
